requirements.txt
uwsgi
Flask>=3.0.0
Flask-Restx
Flask-Cors
pytz
psycopg2-binary
bcrypt
PyJWT
Flask-JWT-Extended
Flask-SocketIO>=5.0.0
requests
python-dotenv
redis
pymongo
# gevent